Make Number of Frames 32 bit.
This amends earlier commits transforming Number of Frames internally from
16 bit into a 32 bit value.

+++ dcmiod/include/dcmtk/dcmiod/modmultiframefg.h
@@ -1,6 +1,6 @@
 /*
  *
- *  Copyright (C) 2015-2016, Open Connections GmbH
+ *  Copyright (C) 2015-2018, Open Connections GmbH
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  This software and supporting documentation are maintained by
@@ -145,7 +145,7 @@ public:
    *  @param  value Value to be stored
    *  @return EC_Normal if successful, an error code otherwise
    */
-  virtual OFCondition setNumberOfFrames(const Uint16 value);
+  virtual OFCondition setNumberOfFrames(const Uint32 value);
 
   /** Set Representative Frame Number
    *  @param  value Value to be stored
